+++ superset-frontend/plugins/plugin-chart-echarts/src/Funnel/transformProps.ts
@@ -35,7 +35,6 @@ import {
   EchartsFunnelLabelTypeType,
   FunnelChartTransformedProps,
 } from './types';
-import { DEFAULT_LEGEND_FORM_DATA } from '../types';
 import {
   extractGroupbyLabel,
   getChartPadding,
@@ -43,7 +42,7 @@ import {
   sanitizeHtml,
 } from '../utils/series';
 import { defaultGrid, defaultTooltip } from '../defaults';
-import { OpacityEnum } from '../constants';
+import { OpacityEnum, DEFAULT_LEGEND_FORM_DATA } from '../constants';
 
 const percentFormatter = getNumberFormatter(NumberFormats.PERCENT_2_POINT);
