+++ superset-frontend/plugins/plugin-chart-echarts/src/index.ts
@@ -45,7 +45,7 @@ export { default as TimeseriesTransformProps } from './Timeseries/transformProps
 export { default as TreeTransformProps } from './Tree/transformProps';
 export { default as TreemapTransformProps } from './Treemap/transformProps';
 
-export { DEFAULT_FORM_DATA as TimeseriesDefaultFormData } from './Timeseries/types';
+export { DEFAULT_FORM_DATA as TimeseriesDefaultFormData } from './Timeseries/constants';
 
 export * from './types';
